Termination: Use unsigned long long for counter variables; reactivate assume no overflow when termination analysis is activated.

+++ lib/goblint/runtime/include/goblint.h
@@ -7,4 +7,4 @@ void __goblint_assume_join(/* pthread_t thread */); // undeclared argument to av
 void __goblint_split_begin(int exp);
 void __goblint_split_end(int exp);
 
-void __goblint_bounded(int exp);
+void __goblint_bounded(unsigned long long exp);
